+++ src/main/java/com/synectiks/cms/repository/BookRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "book", path = "book")
-public interface BookRepository extends PagingAndSortingRepository<Book, Integer> {
+public interface BookRepository extends PagingAndSortingRepository<Book, Long> {
 }
