+++ src/main/java/com/synectiks/cms/repository/CompanyRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "company", path = "company")
-public interface CompanyRepository extends PagingAndSortingRepository<Company, Integer> {
+public interface CompanyRepository extends PagingAndSortingRepository<Company, Long> {
 }
